+++ src/ui/size.hpp
@@ -1,19 +1,18 @@
 #ifndef SIZE_HPP
 #define SIZE_HPP
 
-#include <_types/_uint16_t.h>
 #include <stdint.h>
 
 class Size1D
 {
 public:
-    uint16_t width;
+    float width;
 
     static const Size1D ZERO;
     static const Size1D INFINITE;
     static const Size1D UNDEFINED;
 
-    Size1D(uint16_t width);
+    Size1D(float width);
     ~Size1D() {}
 
     bool isZero() const;
@@ -25,7 +24,7 @@ public:
     Size1D& operator+=(const Size1D& rhs);
     friend Size1D operator+(Size1D lhs, const Size1D& rhs);
 
-    operator uint16_t() const { return width; }
+    operator float() const { return width; }
 };
 
 class Size
@@ -38,7 +37,7 @@ public:
     static const Size INFINITE;
     static const Size UNDEFINED;
 
-    Size(uint16_t width, uint16_t height);
+    Size(float width, float height);
     ~Size() {}
 
     bool isZero() const;
